refactor: MenuItem 컴포넌트 분리
